Style modal transitions
