+++ src/components/Modal.js
@@ -1,6 +1,9 @@
+/* eslint-disable */
 import React, { Component } from 'react';
 import propTypes from 'prop-types';
 import cx from 'classnames';
+import Fade from './Transitions/Fade';
+import Drop from './Transitions/Drop';
 
 class Modal extends Component {
   static propTypes = {
@@ -9,14 +12,14 @@ class Modal extends Component {
   };
 
   static defaultProps = {
-    show: true,
+    show: false,
     children: null,
   };
 
   constructor(props) {
     super(props);
 
-    this.state = { show: true };
+    this.state = { show: false };
   }
 
   componentWillReceiveProps(newProps) {
@@ -28,14 +31,19 @@ class Modal extends Component {
 
     const modalClasses = cx(
       'modal',
-      { 'modal--is-visible': show },
     );
 
     return (
-      <div className={modalClasses}>
-        <div className="modal__background" />
-        <div className="modal__content">{ children }</div>
-      </div>
+      <Fade className={modalClasses} in={show}>
+        <div>
+          <div className="modal__background" />
+          <div className="modal__content">
+            <Drop in>
+              { children }
+            </Drop>
+          </div>
+        </div>
+      </Fade>
     );
   }
 }
