+++ src/styles/components/_modal.scss
@@ -2,12 +2,12 @@ $component-name: modal;
 
 .#{$component-name} {
   opacity: 0;
-  display: none;
   position: fixed;
   top: 0;
   left: 0;
   width: 100%;
   height: 100%;
+  // display: none;
 
   &__background {
     position: absolute;
@@ -28,10 +28,5 @@ $component-name: modal;
     justify-content: center;
     align-items: center;
     flex-direction: column;
-  }
-
-  &--is-visible {
-    opacity: 1;
-    display: block;
   }
 }
